@@ -1,8 +1,9 @@
 package com.brtbeacon.sdk.demo.adapter;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.brtbeacon.sdk.BRTBeacon;
 import com.brtbeacon.sdk.demo.R;
